+++ lib/src/parser/EpisodeDetailsPageParser.dart
@@ -13,7 +13,8 @@ class EpisodeDetailsPageParser {
       print('Episode Title ${episodeTitle.trim()}');
 
       var div = html.getElementById(_EpisodePageNames.DIV_P1_ID);
-      var videoUrl = _extractVidePageUrl(div);
+      var videoUrl = _extractVideoPageUrl(div);
+      var refer = _extractTokenFromUrl(videoUrl);
 
 //    var videoUrl = body
 //        .getElementsByClassName(_EpisodePageNames.ACTION_DOWNLOAD_CLASS)[0]
@@ -34,6 +35,7 @@ class EpisodeDetailsPageParser {
         EpisodeDetails.NEXT: _extractEpisodeId(nextElement),
         EpisodeDetails.PREVIOUS: _extractEpisodeId(previous),
         EpisodeDetails.DESCRIPTION: description,
+        EpisodeDetails.REFER: refer,
       };
     } catch (ex) {
       print(
@@ -57,8 +59,10 @@ class EpisodeDetailsPageParser {
         '';
   }
 
-  String _extractVidePageUrl(Element element) =>
+  String _extractVideoPageUrl(Element element) =>
       element.getElementsByTagName('a')[0].attributes['href'];
+
+  String _extractTokenFromUrl(String url) => url.split('&t=')[1];
 }
 
 class _EpisodePageNames {