+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.1.3+1] - 11/11/2019
+* Enable Refer header property support. EpisodeDetails instances now comes with a refer property. 
+That is used to execute the episode.
+
 ## [0.1.2+1] - 24/12/2019
 * Fixing Anime details page parsing handling no resume available.